// src/models/User.ts
import { Schema, model } from "mongoose";
import { Listing } from "./Listing";
import { Review } from "./Review";
import { id } from "./Types";

// TODO: Make schema for User interface

// export const UserModel = mongoose.model("User", userSchema);

export interface User {
  _id?: id; // Generated by Mongoose
  name: string;
  username: string;
  email: string;
  password: string;
  bookings: id[];
  properties: id[];
  reviews: id[];
  userCreatedAt: Date;
}

const userSchema = new Schema({
  name: {
    type: String,
    required: true,
  },
  username: {
    type: String,
    required: true,
    unique: true,
  },
  email: {
    type: String,
    required: true,
    unique: true,
  },
  password: {
    type: String,
    required: true,
  },
  bookings: [
    {
      type: Schema.Types.ObjectId,
      ref: "Booking",
    },
  ],
  properties: [
    {
      type: Schema.Types.ObjectId,
      ref: "Property",
    },
  ],
  reviews: [
    {
      type: Schema.Types.ObjectId,
      ref: "Review",
    },
  ],
  userCreatedAt: {
    type: Date,
    default: Date.now,
  },
});

export const UserModel = model("User", userSchema);